evaluation/pyproject.toml
[tool.poetry]
name = "modular-rag-platform"
version = "0.1.0"
description = "A modular, no-code RAG deployment platform with RAGAS evaluation"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "rag_platform"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-community = "^0.0.10"
langchain-core = "^0.1.0"
langchain-text-splitters = "^0.0.1"
chromadb = "^0.4.22"
ragas = "^0.1.0"
datasets = "^2.14.0"
click = "^8.1.7"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
pandas = "^2.1.4"
numpy = "^1.26.3"
tqdm = "^4.66.1"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.2"
sentence-transformers = "^2.2.2"
rank-bm25 = "^0.2.2"
unstructured = "^0.11.8"
pypdf = "^3.17.4"
mlflow = "^2.9.2"
rich = "^13.7.0"
networkx = "^3.2.0"
scikit-learn = "^1.3.0"
python-louvain = "^0.16"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.1"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
ipykernel = "^6.28.0"

[tool.poetry.scripts]
rag = "rag_platform.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true